Struct reg_file_parser::reg_value_object::RegValueObject

pub struct RegValueObject {
    pub root: String,
    pub parent_key: String,
    pub parent_key_without_root: String,
    pub entry: String,
    pub value: String,
    pub reg_type: String,
}

Represents a Registry Value

Arguments

root - Registry Root HIVE

parent_key - Registry Value parent key, if applicable

parent_key_without_root - Registry Value parent key, without root HIVE

entry - Registry Value name

value - Registry Value data

reg_type - Registry Value type

Fields
root: String

Registry Root HIVE

parent_key: String

Registry Value parent key, if applicable

parent_key_without_root: String

Registry Value parent key, without root HIVE

entry: String

Registry Value Name

value: String

Registry Value Data

reg_type: String

Registry Value Type

Implementations

impl RegValueObject

pub fn new(
&self,
reg_key_name: &'static str,
reg_value_name: &'static str,
reg_value_data: &'static str,
encoding: &'static str
) -> RegValueObject

Returns a RegValueObject given the following parameters:

Arguments

reg_key_name - Name of Registry Key

reg_value_name - Name of Registry Value

reg_value_data - Data of Registry Value

encoding - Encoding of data. Either UTF8 or ANSI

Example

use reg-file-parser::reg_value_object::RegValueObject;
let reg_value = RegValueObject::new("Settings", "BackgroundColor", "#000000", "UTF8");
assert_eq!(reg_value.value, "#000000");

fn get_hive_without_root(&self, skey: &str) -> String

Gets registry hive without the root hive

skey - Registry Key to get root hive of

fn get_hive(&self, skey: &str) -> &'static str

Returns Registry Root Hive

skey - Registry Key to get root hive of

fn get_reg_entry_value(
&self,
text_line: &'static str,
text_encoding: &str
) -> String

Returns the registry value without the type declaration in front.

text_line - The line of text to parse

text_encoding - The line encoding

fn get_reg_entry_type(&self, text_line: &str) -> &'static str

Retrieves the reg value type, parsing the prefix of the value

text_line - The line of text to parse

This function will return “REG_SZ” if a match is not found.

fn set_reg_entry_type(&self, reg_data_type: &str) -> String

Sets the registry entry’s type from the given data string

reg_data_type - Data type. See Types below

Types

hex: REG_BINARY

hex(0): REG_NONE

hex(1): REG_SZ

hex(2): EXPAND_SZ

hex(3): REG_BINARY

hex(4): REG_DWORD

hex(5): REG_DWORD_BIG_ENDIAN - invalid type?

hex(6): REG_LINK

hex(7): REG_MULTI_SZ

hex(8): REG_RESOURCE_LIST

hex(9): REG_FULL_RESOURCE_DESCRIPTOR

hex(a): REG_RESOURCE_REQUIREMENTS_LIST

hex(b): REG_QWORD

fn get_string_representation(
&self,
string_array: Vec<&str>,
encoding: &str
) -> String

Converts the byte arrays (Vec<&str>) into string

Hews close to C# implementation by using string_builder crate

Arguments

string_array - Vec array to convert

encoding - File Encoding