Module reg_file_parser::utils

Select utilities that do not fit in any other module.

Functions

get_encoding
Parses raw registry file content to determine encoding from registry file version.

strip_braces
Strips braces (‘[’ and ‘]’) from a given &str and returns the result as a String

strip_continue_char
Removes the ending backslashes from the given &str and returns the result as a String

strip_leading_chars
Strips specified leading chars from a &str and returns the result as a String